[core] Add register/unregister_role support; move common keys to keys/ folder
