+++ asistente_ladm_col/gui/gui_builder/gui_builder.py
@@ -15,15 +15,14 @@
  ***************************************************************************/
 """
 from qgis.PyQt.QtCore import (QCoreApplication,
-                              QObject,
-                              pyqtSignal)
+                              QObject)
 from qgis.PyQt.QtGui import QIcon
 from qgis.PyQt.QtWidgets import (QMenu,
                                  QPushButton,
                                  QToolBar)
 
 from asistente_ladm_col.config.config_db_supported import ConfigDBsSupported
-from asistente_ladm_col.config.gui.common_keys import *
+from asistente_ladm_col.config.keys.common import *
 from asistente_ladm_col.config.gui.gui_config import GUI_Config
 from asistente_ladm_col.config.ladm_names import LADMNames
 from asistente_ladm_col.gui.gui_builder.role_registry import RoleRegistry
