+++ asistente_ladm_col/lib/model_registry.py
@@ -17,10 +17,10 @@
 """
 from copy import deepcopy
 
-from asistente_ladm_col.config.gui.common_keys import (ROLE_SUPPORTED_MODELS,
-                                                       ROLE_HIDDEN_MODELS,
-                                                       ROLE_CHECKED_MODELS,
-                                                       ROLE_MODEL_ILI2DB_PARAMETERS)
+from asistente_ladm_col.config.keys.common import (ROLE_SUPPORTED_MODELS,
+                                                   ROLE_HIDDEN_MODELS,
+                                                   ROLE_CHECKED_MODELS,
+                                                   ROLE_MODEL_ILI2DB_PARAMETERS)
 from asistente_ladm_col.config.model_config import (MODEL_ALIAS,
                                                     MODEL_IS_SUPPORTED,
                                                     MODEL_SUPPORTED_VERSION,
